requirements.txt
base58==2.1.0
requests==2.26.0
ConfigArgParse==1.5.2
psycopg2-binary==2.9.1
more-itertools==8.8.0
